Split out `LabelledTextBoxWithPopover` for reuse
